requirements.txt
GitPython
optuna
black
isort
